main.py
import argparse
import os
import asyncio
from dotenv import load_dotenv

# 模拟导入各个 Agent
from agents.collector import CollectorAgent
from agents.analyzer import AnalyzerAgent
from agents.validator import ValidatorAgent
from agents.reporter import ReporterAgent

load_dotenv()

async def run_osint_pipeline(target: str):
    print(f"🚀 Starting OSINT Analysis Pipeline for target: {target}")
    
    # 1. Data Collection
    print("📡 [Collector] Gathering data from sources...")
    collector = CollectorAgent()
    raw_data = await collector.gather_intelligence(target)
    print(f"✅ [Collector] Gathered {len(raw_data)} potential intelligence items.")
    
    # 2. Multi-modal Analysis & Reasoning
    print("🧠 [Analyzer] Parsing multi-modal data and reasoning...")
    analyzer = AnalyzerAgent(model="mimo-v2.5-multimodal")
    analyzed_data = await analyzer.process(raw_data)
    print(f"✅ [Analyzer] Identified {len(analyzed_data)} high-value insights.")
    
    # 3. Cross-Validation (Red/Blue Team)
    print("⚔️ [Validator] Initiating Red/Blue team cross-validation...")
    validator = ValidatorAgent()
    verified_intelligence = await validator.cross_validate(analyzed_data)
    print(f"✅ [Validator] Validated {len(verified_intelligence)} confirmed threats.")
    
    # 4. Report Generation
    print("📄 [Reporter] Generating STIX 2.1 compliant report...")
    reporter = ReporterAgent()
    report_path = await reporter.generate_stix_report(verified_intelligence, target)
    print(f"✅ [Reporter] Report generated successfully at: {report_path}")
    
    print("🎉 Pipeline completed successfully.")

if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="OSINT Agent Network CLI")
    parser.add_argument("--target", type=str, required=True, help="Target keyword or APT group to analyze")
    args = parser.parse_args()
    
    asyncio.run(run_osint_pipeline(args.target))
